cleanup comments in header file
